Move get all traces to dominorun
