+++ domino/aisystems/_util.py
@@ -8,9 +8,7 @@ from urllib.parse import urljoin
 import yaml
 
 from ..authentication import get_auth_by_type
-from ._client import client
-from ._constants import (MIN_MLFLOW_VERSION, MIN_DOMINO_VERSION, LARGEST_MAX_RESULTS_PAGE_SIZE, DOMINO_INTERNAL_EVAL_TAG,
-    EVALUATION_TAG_PREFIX)
+from ._constants import MIN_MLFLOW_VERSION, MIN_DOMINO_VERSION, DOMINO_INTERNAL_EVAL_TAG, EVALUATION_TAG_PREFIX
 from ..exceptions import UnsupportedOperationException, InvalidEvaluationLabelException
 from ..http_request_manager import _HttpRequestManager
 
@@ -29,9 +27,6 @@ def _get_domino_version() -> str:
 def validate_label(label: str):
     if not re.match(VALID_LABEL_PATTERN, label):
         raise InvalidEvaluationLabelException(f"label '{label}' may contain only alphanumeric characters, underscores and dashes.")
-
-def get_is_production() -> bool:
-    return os.environ.get("DOMINO_AI_SYSTEM_IS_PROD", "false").lower() == "true"
 
 def build_metric_tag(metric_name: str) -> str:
     return f"{EVALUATION_TAG_PREFIX}.metric.{metric_name}"
@@ -73,32 +68,3 @@ def verify_domino_support():
 
     if not mlflow_supported:
         raise UnsupportedOperationException(f"This code requires you to install mlflow>={MIN_MLFLOW_VERSION}")
-
-def get_all_traces_for_run(experiment_id: str, run_id: str):
-    filter_string = f"metadata.mlflow.sourceRun = '{run_id}' AND tags.{DOMINO_INTERNAL_EVAL_TAG} = 'true'"
-
-    logging.debug(f"Searching for traces with filter: {filter_string}")
-
-    next_page_token = None
-
-    traces = client.search_traces(
-        experiment_ids=[experiment_id],
-        filter_string=filter_string,
-        page_token=next_page_token,
-        max_results=LARGEST_MAX_RESULTS_PAGE_SIZE,
-        order_by=["attributes.timestamp_ms ASC"],
-    )
-    next_page_token = traces.token
-    while next_page_token is not None:
-        next_traces = client.search_traces(
-            experiment_ids=[experiment_id],
-            filter_string=filter_string,
-            page_token=next_page_token,
-            max_results=LARGEST_MAX_RESULTS_PAGE_SIZE,
-            order_by=["attributes.timestamp_ms ASC"],
-        )
-
-        next_page_token = next_traces.token
-        traces.extend(next_traces)
-
-    return traces
