+++ domino/aisystems/tracing/inittracing.py
@@ -3,7 +3,8 @@ import mlflow
 import os
 from typing import Optional
 
-from .._util import get_is_production, verify_domino_support
+from ._util import get_is_production
+from .._util import verify_domino_support
 
 global _active_prod_model_id
 _active_prod_model_id = None
